+++ .cargo/config
@@ -1,3 +1,3 @@
 [alias]
-unit-test = "test --package voyager_client --bin voyager_client -- test --nocapture"
+unit-test = "test --test unit_test"
 doc-test = "test --doc"
